requirements.txt
curl_cffi
pandas>=2.0.0
openpyxl>=3.1.0
pywebview>=4.0.0
pytz>=2023.3
pywin32>=306
pywin32
bs4
pytest
pytest-mock
requests-mock
usaddress
